refactor: Migrated codebase to flow-based
