+++ src/lib/6.mocks/adapters/sse.mock.ts
@@ -1,4 +1,4 @@
-import { ForSSE } from 'lib/sdk.index';
+import { ForSSE } from 'lib/2.app/driving-ports/for-sse.port';
 
 export const mockSSE = (): jest.Mocked<ForSSE> => ({
   watchRemoteBuilds: jest.fn(),
